requirements.txt
Pillow
scikit-image
pytest
numpy
PySide6
